Add Python bindings for CHLO/StableHLO
In this pull request, I took the existing Python bindings for MLIR-HLO
and mostly copy/pasted them into the StableHLO repository.

One change from MLIR-HLO is that we have two separate bindings for
CHLO and StableHLO in the StableHLO repository. This follows the recent
trend in both repositories where we severed the dependency from CHLO
to MHLO.