Add type inference for stablehlo.reduce (#50)
The new `ReduceOp::inferReturnTypeComponents()` replaces the old `ReduceOp::verify()` with all the checking logic.

Note that a unit test is removed: it had a `llvm.add` inside its region, and intended to return a scale from its region and catch the error in `ReduceOp::verify()`. Now when we replace verifier with infer function, there will be two errors:
1. `'llvm.add' op operand #0 must be integer or LLVM dialect-compatible vector of integer, but got 'f32'`: can be fixed by change `f32` -> `i32`
2. In the last line of region, stablehlo.return supports return tensor only, instead of scalar: No way to fix this.

The HEAD code tolerated these errors because the `ReduceOp::verify()` fails first. Now decide to remove this unit test which is a case that will not happen.