Add type inference for stablehlo.reduce_window (#91)
The new `ReduceWindowOp::inferReturnTypeComponents()` replaces the old `ReduceWindowOp::verify()` with all the checking logic. And expect to cover all XLA checks as well.

And invalid test with `llvm.add` is removed in the same reason as the description of [the PR of "Type inference of ReduceOp"](https://github.com/openxla/stablehlo/pull/50). Please see the reason there.

`ReduceWindowOp` has similar but not same code as `ReduceOp` which is not ideal. So align them to be same in P1~P2 checks.

To address the legacy TODOs, `convertDenseIntAttr` add check for 1-D tensor, and change all caller respectively. Those ops like `SelectAndScatterOp` will be fully addressed in later PRs recently.