Fix CMake build
StablehloOps.cpp references `DataLayout::closest()` and
`DataLayout::getTypeSizeInBits()` but the dependency on
MLIRDataLayoutInterfaces is not mentioned in the CMake rules for the
StablehloOps target.  This patch fixes the problem to prevent an
undefined reference error.  This patch also removes the dependency from
the stablehlo-interpreter target, since it's not the binary (but the
StablehloOps library) that depends on MLIRDataLayoutInterfaces.

This patch also adds `target_include_directories()` for the
`StablehloOps` and `ChloOps` so that path to the Stablehlo header files
is added to the include path of downstream targets.