Populate progress of verifier and type inference. (#168)
The new values in Markdown table of all 114 ops are copied from StableHLO audit sheets with local handy script and the rule to map the old value in sheets to the new value in Markdown table is:
``` 
convert_value_for_md = {
    'Yes (Declarative)': 'yes*',
    'Yes (Imperative)': 'yes*',
    'Yes (need revisit)': 'yes(need-revisit)',
    'TODO': 'no',
    'Yes (revisit)': 'yes(need-revisit)',
    'No (Infeasible)': 'infeasible',
    'WIP': 'no',
}
```
The values of `Type Inference` column is clear and fully up to date. The `Verification` column actually means both ODS and `verify()`. As `verify()` are removed for a lot of ops, the value of them are still set to `yes*`. 
If an op's type inference is done or revisited in the future: its value of `Verification` column will also be revisited and updated at the same time (even if it's set to `yes*`) now.
The sum number in the column title includes all kinds of `yes*` and `infeasible`.