Generalize Element mapping logic
Recent pull requests #162 and #179 prompt the generalization of the
currently ad-hoc "if integer, do this, if float do that, etc" logic
employed in Element's + and sine.

This PR proposes a name for this concept (`map`, to follow the existing
conventions in functional languages), a design and implementation for
it (inspired by `std::visit`, `llvm::make_visitor`, `llvm::TypeSwitch`)
and a refactoring to use it.

For now, let's keep these functions internal to Element.cpp, but if
they end up being required by Ops.cpp, we can make them part of
Element.h.