Expand the definition of the tensor type
In #167, #170 and #172, we are running into the limitations of the
current definition of the tensor type which roughly amounts to
"n-dimensional array with shape S and element type E".

This definition was fine for elementwise op, where only the element type
matters, but for ops like `reshape` and `transpose`, we need to
elaborate the logical representation of a tensor.