Introduce StableHLO evolution process
HLO/MHLO are supported by a wide variety of ML frameworks and compilers,
including IREE, JAX, ONNX, PyTorch, TensorFlow and XLA. With StableHLO,
we are aiming to build on this success and create an amazing portability
layer between ML frameworks and ML compilers.

To that end, we are establishing the StableHLO evolution process to
provide StableHLO users with well-defined means of following and
influencing the evolution of the StableHLO opset and
accompanying infrastructure (the CHLO opset, bytecode
serialization/deserialization, etc).