Add type inference for stablehlo.while (#207)
`.td` file: remove `HLO_PairwiseSameOperandAndResultType`: now we get
the inferred type and compare it against the result type, with the
support of dynamism. There is no need to maintain this trait and add the
dynamism support for it.

`cpp` file: With the power of `isCompatibleForHloTypeInference()`, the
PR optimizes the existing verification logic for `stablehlo.while`: it
can check the 1) count and 2) each pair of type at once for two
TypeRange, so that the code can become more concise.

Tests: The old tests are comprehensive but in wrong places:
1. Negative tests are in `print_while_invalid.mlir` which looks like a
refactor error. This PR renames the file to `verify_while.mlir`.
2. For alignment, move the positive tests from `ops_stablehlo.mlir` to
`verify_while.mlir` well.
3. Add tests for dynamism and unranked cases.