Fix lit checks with ops having implementation defined precision (#265)
The current mode of testing the implementation is to use LLVM's lit-test
framework to syntactically match the output of the operation with a
(user provided) golden output. We do have some ops (like sine, cosine,
tanh) which might produce results with implementation defined precision,
failing the lit checks.

To avoid that this PR allows the lit checks to match the accuracy up to
12 places after the decimal point. The number 12 is chosen arbitrary but
provides a descent approximation of the checked values.