Unconditionally update `description` in op's ODS (#274)
The current version of the checklist recommends to update `description`
only if the implementation of the op conforms to the spec.

When I wrote that recommendation, it was deliberate. I thought that the
process could be: 1) if the implementation doesn't conform don't update
the description, 2) fix the tickets, 3) update the description.

But now I think that this simply introduces an extra step into the
process (step number 3) without a good reason. If there's a spec for the
op, it's already useful to make it part of documentation. And if there
are some bugs in the implementation, that's fine as long as we have
tickets for them.