Move shape function of TransposeOp to shared file (#322)
This PR moves the shape function of TransposeOp including the bounds to
shared files. This is an example how we add bounds and move code for
other ops in the near future.

Question: Does the existing code
https://github.com/openxla/stablehlo/blob/main/stablehlo/dialect/StablehloOps.cpp#L4189
mean that we can't have bounds and sparse tensor encoding at the same
time? (Eugene: Correct.)

The reason we prefer `inferReturnTypes` over
`inferReturnTypeComponents`:
inferReturnTypeComponents does not work: this assert block it:

[inferReturnTypeComponents](https://github.com/llvm/llvm-project/blob/main/mlir/lib/Interfaces/InferTypeOpInterface.cpp#L190)
So I will keep using current inferReturnTypes

@smit-hinsu PTAL, thanks!