Refactor pretty printing machinery to separate files (NFC) (#332)
Created `stablehlo/dialect/AssemblyFormat.h` with shared code for pretty
printing. This only depends on `StablehloBase` and upstream MLIR.

Skipping the move of print operations that rely on specific attribute
types. Will reconsider these in a later PR. Custom printer machinery
requires functions to have a certain signature, so there is a good
change these will need to be templated.
`StablehloOps.cpp` has many `using mlir::hlo::print*` declarations. An
alternative is to `using namespace mlir::hlo` in the file scope, but
that is not super desirable and clang tidy complains.