Allow inlining StableHLO ops
This hasn't been discussed much, but as part of defining the
relationship between StableHLO and MHLO, it looks reasonable to avoid
doing transformations on StableHLO and leave that to MHLO.

This way, we don't have to duplicate stuff like constant folding,
canonicalization, etc etc.

Inlining is one of the transformations that's in a gray area because
this is something that we get for free from upstream, so I don't think
that the duplication argument applies here.