Prettyprint CustomCallOp target using SymbolRef printing (#432)
This is consistent with `func.call` printing, and relies entirely on
upstream machinery. Using this printing is cleaner and more consistent,
and if we eventually consider modifying what is permitted in
CustomCallOp target names, this printing will not need to be updated.

```
%0 = stablehlo.custom_call @foo(%arg0) ...
%1 = stablehlo.custom_call @"foo-not-id"(%arg0) ...
```

This is a little messier for non-identifier target names, but there are
_very_ few cases that target non-C or non-MLIR identifier names. The
only ones I was able to find were in a single test file in mlir-hlo.

The implementation directly calls upstreams
`AsmPrinter::printSymbolName`/`AsmParser::parseSymbolName`. These
methods are able to handle identifier and non-identifier targets.

Closes #431