Use new github actions output environment files (#459)
GitHub Actions is deprecating save-state and set-output commands, need
to update to use new Environment Files:

https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

I made a PR on my forked repo to test, and it had the following in the
logs:

```
Run ./.github/actions/setup-build
  with:
    llvm-version: d4c61314420ca8794ad6d2588566c64109482505
  env:
    LLVM_PROJECT_DIR: llvm-project
    LLVM_BUILD_DIR: llvm-build
    STABLEHLO_BUILD_DIR: stablehlo-build
    STABLEHLO_PYTHON_BUILD_DIR: stablehlo-python-build
```

This confirms that the new syntax is working. The `llvm-version` is set
by the output of the previous command.