(abs..constant) Align examples in spec/ODS with spec_checklist.md (#453)
In #439, we documented the conventions for examples in spec/ODS, and I
went to check how closely the previous work on the spec follows these
conventions.

I discovered quite a few divergencies, so I started going through the
spec fixing things. This PR covers 10 out of the currently specced 43
ops.