Build StableHLO as a part of LLVM in CI (#457)
A suggestion made in #444, build StableHLO as a part of LLVM in CI. Many
clients do this so we should make sure the build is clean.

Other minor cleanup: Remove matrix, we only run tests on ubuntu
currently. No need for a matrix, no need to build for anything except
host.