update spec checklist: Adding ticket link in PR (#570)
For tickets like https://github.com/openxla/stablehlo/issues/369 which
is applicable to many PRs and prevents the status of one or more
implementations from being a `yes`, it would be useful to track all the
Ops which needs to be tracked before closing this issue. Linking the
ticket in the spec could be a solution and hence the proposal.