Add spec for parallel execution semantics (#634)
I think that recent PRs with specs for collective ops (e.g. #503) would
benefit from a more formal definition of parallel execution semantics of
StableHLO programs.

That will provide additional clarity which is needed at the moment (see
discussions in the PRs) and reduce the duplication between the specs of
individual collective ops.

The description of execution semantics isn't yet as formal as I'd like,
but it can already benefit specs for collective ops. I'll leave the rest
to follow-up work within #484.