Change -1 to ? printing when possible (#569)
Prepare for dynamic dim size to be changed from -1 to INTMIN.

This is a first step at removing the reliance on upstream
`kDynamicSize`. There is more refactoring that can be done, but this
change will at least make it so that some of our error messages will
display `?` instead of the value, and our infer tests now print "?" as a string for consistency.