Fix verifier check disallowing dynamic shapes in reduce and related ops (#671)
In the verification of the reducer function, while verifying that the
block argument shape is a sub-sequence of the computed output shape, it
doesn't consider dynamic shapes.

The original intention here might have been to check that
`allowedDimensions[argShapeIdx]` was dynamic and not
`argShape[argShapeIdx]`. Keeping the original check for now but I wonder
if the op should simply disallow dynamic block arguments. Those are not
going to be refined based on the input shapes. If so, such a check could
be added later on.