Improve type inference for SelectOp (#692)
The implementation says "the output shape should be the most general of
the operand shapes at each dimensions", but falls a bit short of
delivering that. This PR fixes it.

Also merges the verifier into the shape function, to follow the design
pattern that we've established for type inference, and switches to
`inferMostSpecificType` which is used for the same purpose in several
other ops.