Remove region verification in inferReduceWindowOp
TypeInference should not be verifying as a goal as that is handled by
the verifier. Additionally, this is removed because it made type
inference useless, as type inference required a region when building the
op, but the op does not have a builder that also constructs or takes a
region.