Add spec for ConvolutionOp (#441)
fixes #359 

The PR addresses the followings:
1. Spec of ConvolutionOp
2. Clarify the semantics of `precision_config` : The precision_config
parameter is a array of enums without any constraint on its size. Need
to resolve this.
- update: Added constraints on the parameter. With that the verifier is
in sync with this spec. Also added
https://github.com/openxla/stablehlo/issues/445 for further exploration.
4. Fix
https://github.com/openxla/stablehlo/pull/360#issuecomment-1282808168
5. Avoid disabling clang formatting in StablehloOps.cpp.
6. Address https://github.com/openxla/stablehlo/issues/399

Only missing peice:

The constraint between output feature size and input batch size. Working
on getting a better understanding on this: Done

Type inference should be "revisit" as well because of #600.