Temporarily allow empty `replica_group` for MHLO parity (#767)
[Here](https://github.com/openxla/stablehlo/issues/498#issuecomment-1331396793)
we decided on StableHLO to dis-allowing empty `replica_groups` but that
needs some extra work to achieve that.

If we want to share
[verifyReplicaGroups](https://github.com/openxla/stablehlo/blob/59b68c586ea1381a7cb53e1217904f31755bb390/stablehlo/dialect/TypeInference.cpp#L326)
logic with MHLO, we cannot just prevent empty `replica_groups` as there
exists MHLO based tests
[allowing](https://github.com/tensorflow/tensorflow/blob/1c83485345d1d7ff3e9c3afcce55d303e793e216/tensorflow/compiler/xla/mlir_hlo/mhlo/IR/hlo_ops_common.h#L67)
that.

The current PR will temporarily preserve the same behavior of MHLO to
allow code sharing till #498 is fixed. Note the verification for all the
collective ops are already marked as `revisit`.