Fix crashing builders for CaseOp/IfOp/MapOp (#760)
Shape function can be called directly from autogenerated `build()`
function, which may not guarantee the added region(s) in
`odsState.regions` to be non-empty. Need check it here to avoid a crash
for the ops that need regions in type inference, like
`IfOp/CaseOp/MapOp`.

An alternative way is to consider disable the autogenerated `build()` at
all: so far we have no strong evidence that this is a better way and
have not find a solution to this approach. While, This PR avoids
surprised crash in dev of these ops and gives the error instead.

Note: in the tests of `ops_stablehlo.mlir`, a ODS-valid op must have at
least 1 block in each region: so that this error will not trigger in the
tests.

Doc is updated separately in
https://github.com/openxla/stablehlo/pull/763