Add support for f8E4M3FN and f8E5M2 (#774)
The FP8 RFC has been approved
(https://github.com/openxla/xla/discussions/22), so I'm following up and
adding support for the proposed FP8 data types to StableHLO.

As discussed in RFC comments, there are other FP8 data types
(https://github.com/openxla/xla/discussions/22#discussioncomment-4337622),
and they can be included in StableHLO as well in the future, following
the same RFC process. Please let us know if you're interested in working
on this, and we will be happy to help!