Add versions to attributes, enums, and types (#736)
- Moved `TokenType` to ODS in VHLO to add consistent implementation for
versioned interface. We can consider moving StableHLO/MHLO TokenType
decls to ODS as well.
- Added a `Versioned(Op|Attr|Type)Interface` declaration to version
ops/attrs/types.
- Moved `Vhlo(Enums|Types|Attrs).h.inc` to bottom of `VhloOps.cpp` since
the version interfaces had a dependency on
`VhloDialect::getCurrentVersion()`.
- Updated legality checks to include checking attr and type versions
satisfy the target version.
- Added test point for legal types/attrs, currently only
`OutputOperandAlias`, added in `0.4.0` is the only attribute that can be
tested for illegality.

Currently, there are only `V1` attributes and types. This means we do
not need to worry about type or attribute converters in the current
StableHLO release. To avoid introducing technical debt for the sake of
testing type/attr conversion, I think it is best we submit conversion
code after we have a use case for it.

I have confirmed that Type and Attribute conversions are possible, and
have prototyped a few machineries for these conversions. See:
https://github.com/openxla/stablehlo/commit/334b88ba190fc8127e76e60f2072a7b0f5f2ef2b

Closes #676.