Use AnyType and AnyAttr for all types and attributes in VHLO (#765)
VHLO is intended to be a representation of the IR format, not an exact
typed dialect, as such use AnyType/AnyAttr everywhere. This change makes
use of AnyType for all types.

The definition of AnyType can be updated to only include VHLO types,
`AnyVhloType`, but this isn't necessary and should follow the work
regarding forked types and attrs (#674).

Could not use `AnyAttr` for `DefaultValued` fields since the constructor
of the attribute is required to accept the default value, so there are
some remnants of `DefaultValuedStringAttr`. I also did not update
`Shape_WitnessType` since that type needs to be forked. That update will
come in the forking types PR.

Closes #734.

Note this PR is built on top of versioned attributes/types (#736) and
should be reviewed/submitted after. Can [use this
diffbase](https://github.com/openxla/stablehlo/pull/765/files/cee7cd6ca7c96525239d51e199a079348054d221..0c0eb99b94bbd1d5883c1fbaeb83617564aa96d9)
or wait until other is merged to review.