Move 26 verifiers shape functions to shared target (#783)
To continue deduplicate the verifiers and shape functions among
StableHLO and MHLO, this PR, following
https://github.com/openxla/stablehlo/pull/26, continue move 24 verifiers
& shape functions from `StablehloOps.cc` to `TypeInference.h/cpp` with
change to static methods and organize with alphabetical order.

This PR also includes:
1. Unify emit error methods with static `emitOptionalError()`
2. Similar to the discussion of
https://github.com/openxla/stablehlo/issues/481, in the signuature of
the new `inferAllToAllOp()`, I use `int64_t` instead of `uint64_t` to
pass all the parameters to avoid verbose "static_cast to int" for each
of attr in the method body.
5. **Few or trivial tests changes**: No semantic changes or
optimization.

Includes moves for (they are select in the first half of code for
`StablehloOps.cc`) : strikethrough means the verifiers are merged into
corresponding shape functions.
1. inferAbsOp
1. inferAllToAllOp
1. inferBroadcastOp
1. inferCholeskyOp
1. inferClampOp
1. inferComplexOp
1. inferConstant
1. inferDynamicSliceOp
1. inferIsFiniteOp
1. inferRealOp
1. inferGetTupleElementOp
1. inferTupleOp
1. verifyAllReduceOp
1. verifyBitcastConvertOp
1. ~~verifyBroadcastOp~~
1. verifyBroadcastInDimOp
1. ~~verifyClampOp~~
1. verifyCollectivePermuteOp
1. verifyDynamicBroadcastInDimOp
1. verifyDynamicReshapeOp
1. ~~verifyDynamicSliceOp~~
1. ~~verifyGetTupleElementOp~~
1. verifyIotaOp
1. verifyRealDynamicSliceOp
1. verifyReduceScatter
1. ~~verifyTupleOp~~

Expect there are ~30 methods remaining to be moved. Why not include them
all before send this PR out? That will need more time and it's better to
send PRs early and frequently to avoid code conflict burden.