Add shell script and GitHub Action to run markdownlint (#814)
As per #792, here's a script to run `markdownlint`. The bash script
includes usage instructions. You just need to pass it a file, directory,
or glob pattern for the MD files you want to lint. It uses a Docker
image because I fear the npm package version dependency hell, but the
docker image is versioned.