First step towards turning spec.md into a programming language spec
As the first step towards implementing #484, this PR introduces an
EBNF grammar for the StableHLO language, modelling it after the generic
assembly format of the StableHLO dialect.

In addition to creating a foundation for future work as described in
 #484, this PR immediately improves clarity of the Types and Constants
sections of the specification.

Next steps involve: 1) writing the Notation section along the lines of
 #289, 2) formalizing execution of the StableHLO language in the
Execution section.