Replace "foo of any supported type" with "foo"
This is a convention that dates back to the very first specs that we
wrote, but upon second look I think it's just boilerplate. E.g. we don't
say "function of any supported types" or "tuple of any supported types".

The newly introduced formalization for types makes it very clear what
are the various supported tensor types, value types, etc, so let's make
things easier to read.