Consistently use HLO_StaticShapeTensor
Some ops in the StableHLO dialect have "load-bearing" result shapes,
meaning that these shapes cannot be inferred from op's operands,
attributes and regions.

For these ops, the current design philosophy involves one of the two
approaches:
  1) The shape is specified statically in the result type, e.g. see
     ReshapeOp.
  2) The shape is passed dynamically as an operand (or can be computed
     indirectly from shape-related operands), e.g. see DynamicReshapeOp.

In the future (#8), we want to have one unified approach in these
situations, but we're not there yet, and this PR is about an incremental
improvement to the state of the art.

One important piece of the design philosophy for 1) is that these ops
need to be restricted to produce HLO_StaticShapeTensor, otherwise it is
unclear what their semantics is.

  %1 = "stablehlo.reshape"(%0) : (tensor<2x4xf32>) -> tensor<?x?xf32>

For example, the code snippet above doesn't make sense because we need
an actual shape to execute the op - either passed at compile time or
passed at runtime - and neither is happening here.

When working on #622, I discovered that there are some ops which fall
into 1) but don't have their result types restricted to be static:
InfeedOp, IotaOp, RecvOp and RngBitGeneratorOp.

This means that StableHLO producers can currently produce these ops
with dynamic result types, which would be unsound. This PR fixes this
issue by changing the result types to HLO_StaticShapeTensor or
equivalent.