Make CHLO ops inlinable
When working on #622, I discovered that functions with CHLO ops aren't
getting inlined, which is a miss because there's no reason for these ops
to not be inlinable. For comparison, StableHLO and MHLO both implement
this interface.

This PR fixes that and adds tests to make sure that things stay this way
in the future as well.