go back to using no indentation for lists
This ensures simplicity and clarity of the rules, especially when
comparing ordered and unordered lists, because there's no lint check
for indentation of ol lists, and no objective reason to indent them.
This will hopefully avoid confusion about how much indentation is
needed, but most of all, it saves chars on line length.