Handle bounds in the SelectAndScatterOp shape function (#917)
SelectAndScatterOp's bound inference is very simple. There are two
inputs: `operand` and `source`. The `source` contains the indices to the
`result` tensor to place the reduced values from `operand`, so it does
not affect the bound of the `result` tensor. As for `operand`, the type
is equal to the `result` tensor. There is only a movement of values
happening in this op (after reduction is applied, but this also does not
affect the bounds), so we can propagate the bounds with no further
inference done.

closes #754