Handle bounds of Gather op (#908)
The dimension sizes of result, corresponding to offset dimensions,
depend on attributes (like `collapsed_slice_dims` and `slice_sizes`) and
hence are always static.

Whereas, the dimension sizes of result, corresponding to batch
dimensions, depends on input `start_indices` and could be dynamic. The
corresponding bounds, in that case, are propagated from the
`start_indices`.



Side notes:
I deliberately avoid making any changes in `inferGatherShape` which is
currently shared by `GatherOp::inferReturnTypeComponents` and
`GatherOp::reifyReturnTypeShapes`. The reason being I do not
`reifyReturnTypeShapes` needs to handle the bounds.

Because of the decision, we have some duplicate code in line 1237
(duplicate of
https://github.com/openxla/stablehlo/blob/afcbb953c12a2a3a8564cab73278f4b47021ec01/stablehlo/dialect/TypeInference.cpp#L1149).

PTAL @smit-hinsu