Handle bounds in `TriangularSolveOp` shape function (#930)
The shape and bounds of batching dimensions is inferred using
inferMergedDimAndBound.

If non-batching dimensions of both A and B are dynamic with bound, then
the following inference rule applies. The shape is propagated the same
way bounds are propagated. Arbitrary bounds are provided for A, B, and X
to visually aid in how the they are moved:
```
A * X = B, R=rank

                | A   | X   | B   | X[R-2], X[R-1]
---------------------------------------------------------------
left_side=false | 3x3 | 3x5 | 3x5 | min(A[R-1], A[R-2]), B[R-1]

X * A = B, R=rank

                | X   | A   | B   | X[R-2], X[R-1]
---------------------------------------------------------------
left_side=true  | 2x3 | 3x3 | 2x3 | B[R-2], min(A[R-1], A[R-2])
```

closes #807