Remove type inference for ReturnOp
I added a shape function for ReturnOp when working on an initial version
of #622 which was relying on --tf-shape-inference.

TensorFlow's shape inference needs all ops which are not hardcoded
in the pass to have shape functions, so I added one for ReturnOp even
though it is has unclear utility beyond this specific use case (ReturnOp
doesn't have any results, so its shape function returns an empty vector,
which isn't super useful).

Now that we're working towards a dedicated shape refinement pass for
StableHLO, we no longer need a shape function for ReturnOp, so I'm
proposing to remove it in this PR.