Enable specializing dynamically-shaped StableHLO programs to static shapes (#1041)
Dynamic shapes in StableHLO are an appealing feature for ML frameworks,
but not all ML compilers implement them. As a result, there is a
practical need for infrastructure that would bridge this gap by
specializing dynamically-shaped StableHLO programs to static shapes.

For example, we may want to take the following StableHLO program that
performs a broadcasting addition of two dynamically-shaped tensors and
then specialize it to static shapes based on ahead-of-time or
just-in-time knowledge of the static shapes of its arguments.

```
func.func @main(
    %arg0: tensor<i32>,
    %arg1: tensor<?x4xf32>,
    %arg2: tensor<2x?x4xf32>) -> tensor<2x?x4xf32> {
  %0 = stablehlo.constant dense<2> : tensor<1xi32>
  %1 = stablehlo.reshape %arg0 : (tensor<i32>) -> tensor<1xi32>
  %2 = stablehlo.constant dense<4> : tensor<1xi32>
  %3 = stablehlo.concatenate %0, %1, %2, dim = 0 : (tensor<1xi32>, tensor<1xi32>, tensor<1xi32>) -> tensor<3xi32>
  %4 = stablehlo.dynamic_broadcast_in_dim %arg1, %3, dims = [1, 2] : (tensor<?x4xf32>, tensor<3xi32>) -> tensor<2x?x4xf32>
  %5 = stablehlo.add %4, %arg2 : tensor<2x?x4xf32>
  return %5 : tensor<2x?x4xf32>
}
```

Let's say that we know that %arg0 (which carries the size of the
unbounded dimensions of %arg1 and %arg2) is equal to 3, then the
corresponding statically-shaped program might look as follows.

```
func.func @main(
    %arg1: tensor<3x4xf32>,
    %arg2: tensor<2x3x4xf32>) -> tensor<2x3x4xf32> {
  %0 = stablehlo.broadcast_in_dim %arg1, dims = [1, 2] : (tensor<3x4xf32>) -> tensor<2x3x4xf32>
  %1 = stablehlo.add %0, %arg2 : tensor<2x3x4xf32>
  return %1 : tensor<2x3x4xf32>
}
```

This PR implements an initial version of --stablehlo-refine-shapes, a
pass that propagates shape information across StableHLO programs. In
this initial version, we're making the following simplifying
assumptions:

(A1) Current design for StableHLO dynamism. We are using the current
design for StableHLO dynamism, but we're also planning to update this
pass to account for the changes in the upcoming dynamism RFC when it's
ready. We found it useful to implement the first version of shape
refinement prior to implementing the new design for dynamism and not the
other way around in order to speed up the delivery.

(A2) Shape polymorphism. We are limiting ourselves to StableHLO programs
that support shape polymorphism. In such programs, only data can depend
on shapes (e.g. a broadcast can produce data of a dynamically-computed
shape), but shapes cannot depend on data. Therefore, the shapes of all
intermediate values can be computed from the shapes of the inputs.

(A3) Full specialization. We are only interested in full specialization
to statically-shaped StableHLO programs. In alignment with A2, this
means that we want the resulting programs to have statically-shaped
types and constant shape values. Specialization to bounded dynamism
might be an interesting avenue for future work.

(A4) Inlined programs. We assume that these programs consist of a single
StableHLO function, i.e. that it's fine to ask users to inline their
programs into one main function before running shape refinement.