Move VhloTypeDefs.cpp.inc into VhloTypes.cpp (#1053)
As discussed in #899, it is desirable to move the following includes
from VhloOps.cpp to VhloTypes.cpp:

  #include "stablehlo/dialect/VhloTypeInterfaces.cpp.inc"
  #define GET_TYPEDEF_CLASSES
  #include "stablehlo/dialect/VhloTypeDefs.cpp.inc"

This way, vhlo_types becomes self-contained, i.e. we will no longer have
to link its users with vhlo_ops. Given the fairly involved dependency
structure around the VHLO dialect, this is a desirable design property.