Link frontend contract tickets in the spec (#1076)
Now that we have clearly identified the deltas between how the StableHLO
dialect is used by frontends and what is specced (see
https://github.com/orgs/openxla/projects/6), I figured it would be good
to link these deltas in the spec, so that it's clear what areas are
expected to change in the near future.

As I was working on this pull request, I also called out several tickets
which are not part of the Frontend Contract project, because they
represented highly likely todos from the New Features project. I think
this is also useful to include in the spec for the readers to understand
the anticipated direction of StableHLO evolution.