Fork FuncOp and CallOp to VHLO (#899)
Leaving out `ModuleOp` for now. It is awkward to convert ModuleOp to
something else, also would also require adding `--no-implicit-module` to
all calls which is less ergonomic and easy to miss. May decide to avoid
forking `ModuleOp`, and instead just enforce that the op has valid
types/attrs. If that op disappears we can get clever in the
BytecodeReader. We can discuss this further next week.

This PR (built on top of ):
- Fork Ops: `FuncOp, CallOp`
- Fork Types: `FunctionType`
- Fork Attributes: `DictionaryAttr, TypeAttr` (`TypeAttr` used for
`FunctionType` attribute)
- Add conversions: 
  + `func.func` <--> `vhlo.func`
  + `func.call` <--> `vhlo.call`
  + `func.return` --> `vhlo.return`
+ `vhlo.return` --> `func.return` if parens is `vhlo.func`, else
`stablehlo.return`
- Added pretty print for func ops, since these are so common and generic
format of functions is difficult to write tests for.

Closes #850.