Implement Region evaluation (#1093)
Closes #995 

The implementation follows the discussion in #995 in the following ways:

## Class `InterpreterScope`

To represent runtime stack corresponding to a region of a program under
evaluation. The idea is to use a "parent-child" hierarchy of the class
objects to implement the scoping rules.
1. Name : `InterpreterScope`. Refrained fro using just `Scope` because
that seems more like a static program concept.
 2. Members variables:
- `stackFrame`: mapping from SSA values, defined in the current region,
to their evaluated runtime `Tensor` values.
- `parentScope`: a handle to `InterpreterScope` object corresponding to
the (syntactically) enclosing region.
 3. Member functions:
- `void add(Value ssaValue, Tensor runtimeValue)`: Add the mapping SSA
value (`ssaValue`), defined in region, to its runtime value
(`runtimeValue`) in the current scope.
- `Tensor find(Value ssaValue) const;`: Find the runtime value mapped to
SSA value `ssaValue`. The search starts
with the current scope and then recursively continues over to the scope
defined by `parentScope`.
 
A few points to note:
1. A `InterpreterScope` object is instantiated every time a region is
evaluated. This is just a simplification over the following alternative
- Store a map from a ` region` to corresponding `InterpreterScope`
object.
- Whenever a region is evaluated used the mapped `InterpreterScope`
object for scope resolution.

2. A `InterpreterScope` object treats the `parentScope` as immutable to
align with the fact that a StableHLO program, in pure SSA form (without
memory allocation/load/store ops), disallows mutating the `parentScope`
object from within a region.

3. Copy prohibited for `InterpreterScope` objects.

## Region evaluation

The evaluation of a region is realized using

```
/// Evaluates an mlir::Region `region` using the runtime values `runtimeArgs`
/// corresponding to the arguments of the containing block, assuming that the
/// region has only one block. Interprets the operations within the block and
/// returns the runtime values for the terminator's arguments.
llvm::Expected<llvm::SmallVector<Tensor>> eval(Region &region, llvm::ArrayRef<Tensor> runtimeArgs,
  const InterpreterScope *const parentScope);
```

The above function is used to evaluated the region for any op including
the `func::FuncOp`. Its definition provides a sole place when a
`InterpreterScope` is stack-allocated and its `parentScope` is
initialed.

### How the `eval(Region..` is invoked

The evaluator of  `func::FuncOp` invokes the function as
```
eval(func.getBody(), args, nullptr); // This is because IsIsolatedFromAbove trait for func::FucnOp is true
``` 

The evaluator for other StableHLO ops, say `stableHLO::whileOp`, invokes
the function as
```
bool isIsolatedFromAbove = cond.getParentOp()->hasTrait<OpTrait::IsIsolatedFromAbove>();
eval(cond, runtimeInputs, isIsolatedFromAbove ? nullptr : &scope);
```
Note: 
- `scope` is the `InterpreterScope` object corresponding to the region
containing the `whileOp`
- As the evaluator for an op expected to know the semantics, one might
get tempted to skip the check for
[OpTrait::IsolatedFromAbove](https://mlir.llvm.org/docs/Traits/#isolatedfromabove)
and directly pass the appropriate scope. Please note that these checks
make the code adjust to any future changes in the behavior of the op
w.r.t its visibility to outer scope.
- The way the `parentScope` is initialized makes sure the following
relationship to hold between a `InterpreterScope` object and its stored
`parentScope`
- `InterpreterScope Scope`: Defined whenever a `region` is evaluated
- `Scope.parentScope`: `InterpreterScope` object corresponding to the
syntactically enclosing region

## [Scoping rules](https://mlir.llvm.org/docs/LangRef/#value-scoping) to
cover

1. **A value defined by an argument to a region can always be used by
any operation deeply contained in the region:** Covered by the design of
`InterpreterScope`.
2. **A value defined in a region can never be used outside of the
region**: For a value defined in a region `RI` to be used outside of the
region `RO`, then `RO`, during its evaluation, should receive the
`parentScope` as the `InterpreterScope` object corresponding to `RI` ,
which will never happen per how `parentScope` is always set to
`InterpreterScope` object corresponding to the syntactically enclosing
region. For example,
 
For the evaluation of `R1` and `R2`, `parentScope` is set to
`InterpreterScope` object corresponding to the `R`.
```
  { // enclosing Region, R
    { // region R1
      %value_defined_in_a_region = ...
    }, { // region R0
      %operation_wants_to_use_value_defined_in_a_region = xyz(%value_defined_in_a_region
    }
   // 
```

3. **A value defined in a region can be used by an operation which has a
parent in the same region, if and only if the parent could use the
value**:

Here is the example to demonstrate the above rule
```
   { // Region R
       %value_defined_in_region = ...

       // CASE: I
       %parent_defined_in_same_region_as_value_and_can_use_value = ...({ '' Region R1
            %operation_uses_value = ...(%value_defined_in_region, ...
       }) 

      // CASE : II
      %parent_defined_in_same_region_as_value_and_cannot_use_value_as_isolated_from_above = ...({ // Region R2
            %operation_uses_value = ...(%value_defined_in_region, ...
       })
   }
 
```

Case I: For evaluation `R1`, the `InterpreterScope` object corresponding
to the `R` is provided as `parentScope`. Hence the value
`value_defined_in_region` will be visible.

Case II: For evaluation `R2`, the `nullptr` provided as `parentScope`
depending on the op trait
[OpTrait::IsolatedFromAbove](https://mlir.llvm.org/docs/Traits/#isolatedfromabove).
Hence the value `value_defined_in_region` will not be visible.

## Testing 

As there is no region op currently implemented to test the PR, we have
implemented a simplified version of `whileOp`. The simplification is
that if the runtime value corresponding to `cond` block is `true`, then
we iterate the loop for fixed (2) times.The simplification has to be
made because there is no `compreOp` implemented at time time.

We used the `whileOp` to implement a program, `interpret_while.mlir`,
which captures values from various allowable scope. The illegal
reference to scopes will be flagged by the versifier only. So the
interpreter can assume the input program to always obey the scoping
rules and thereby avoid checking for illegal cases.


update: Changed name `InterpretScope` to `Scope`
[Lexical_scope_vs._dynamic_scope](https://en.wikipedia.org/wiki/Scope_(computer_science)#Lexical_scope_vs._dynamic_scope)